app: ovs-offload: append rss action
Flow rules can be added on behalf of representee using representor
ports. Even if application is configured with multiple queues to support
RSS, UCast rules are installed for these representees via representor ports
because representor ports are single queues. Handling these scenarios by
appending RSS action lo flow action list if application is configured for
multiple queues.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/130818
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>

+++ app/ovs-offload/include/ood_msg_ctrl.h
@@ -134,7 +134,7 @@ typedef struct ood_msg_ack_data1 {
 typedef struct ood_msg_ready_data {
 	uint8_t val;
 	uint16_t nb_ports;
-	uint16_t data[];
+	uint32_t data[];
 } __rte_packed ood_msg_ready_data_t;
 
 /* Exit msg */
